refactor(normalization-core): add canonical toErrorObject and migrate six copies
