+++ packages/media-core/src/read-response-with-limit.ts
@@ -1,4 +1,5 @@
 // Media Core module implements read response with limit behavior.
+import { toErrorObject } from "@openclaw/normalization-core/error-coercion";
 import { resolveTimerTimeoutMs } from "@openclaw/normalization-core/number-coercion";
 
 async function readChunkWithIdleTimeout(
@@ -40,7 +41,7 @@ async function readChunkWithIdleTimeout(
       (err: unknown) => {
         clear();
         if (!timedOut) {
-          reject(toLintErrorObject(err, "Non-Error rejection"));
+          reject(toErrorObject(err, "Non-Error rejection"));
         }
       },
     );
@@ -183,18 +184,4 @@ export async function readResponseTextSnippet(
     return `${collapsed.slice(0, maxChars)}…`;
   }
   return prefix.truncated ? `${collapsed}…` : collapsed;
-}
-
-function toLintErrorObject(value: unknown, fallbackMessage: string): Error {
-  if (value instanceof Error) {
-    return value;
-  }
-  if (typeof value === "string") {
-    return new Error(value);
-  }
-  const error = new Error(fallbackMessage, { cause: value });
-  if ((typeof value === "object" && value !== null) || typeof value === "function") {
-    Object.assign(error, value);
-  }
-  return error;
 }
