+++ packages/memory-host-sdk/src/host/embeddings.ts
@@ -1,4 +1,3 @@
-// Memory Host SDK module implements embeddings behavior.
 import { DEFAULT_LOCAL_MODEL } from "./embedding-defaults.js";
 import { sanitizeAndNormalizeEmbedding } from "./embedding-vectors.js";
 import { createLocalEmbeddingWorkerProvider } from "./embeddings-worker.js";
@@ -9,6 +8,8 @@ import {
   type LlamaEmbeddingContext,
   type LlamaModel,
 } from "./node-llama.js";
+// Memory Host SDK module implements embeddings behavior.
+import { toLintErrorObject } from "./retry-utils.js";
 import { normalizeOptionalString } from "./string-utils.js";
 
 type DisposableResource = {
@@ -202,18 +203,4 @@ export async function createLocalEmbeddingProviderInProcess(
       return closePromise;
     },
   };
-}
-
-function toLintErrorObject(value: unknown, fallbackMessage: string): Error {
-  if (value instanceof Error) {
-    return value;
-  }
-  if (typeof value === "string") {
-    return new Error(value);
-  }
-  const error = new Error(fallbackMessage, { cause: value });
-  if ((typeof value === "object" && value !== null) || typeof value === "function") {
-    Object.assign(error, value);
-  }
-  return error;
 }
