+++ packages/normalization-core/src/error-coercion.ts
@@ -1,0 +1,18 @@
+/**
+ * Normalizes an unknown thrown value into an Error. Non-Error objects become
+ * the `cause` and have their enumerable fields copied so structured details
+ * (codes, statuses) survive the coercion.
+ */
+export function toErrorObject(value: unknown, fallbackMessage: string): Error {
+  if (value instanceof Error) {
+    return value;
+  }
+  if (typeof value === "string") {
+    return new Error(value);
+  }
+  const error = new Error(fallbackMessage, { cause: value });
+  if ((typeof value === "object" && value !== null) || typeof value === "function") {
+    Object.assign(error, value);
+  }
+  return error;
+}
