+++ packages/normalization-core/src/index.ts
@@ -1,6 +1,7 @@
 // Public barrel for shared coercion and normalization helpers.
 
 export * from "./boolean-coercion.js";
+export * from "./error-coercion.js";
 export * from "./number-coercion.js";
 export * from "./record-coerce.js";
 export * from "./string-coerce.js";
